+++ Changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 Version 0.4.4
 -------------
+* added coordinates functions
 
 
 Version 0.4.3